Lots of design improvements
