+++ discord_widget.php
@@ -1,0 +1,1 @@
+<iframe src="https://discordapp.com/widget?id=687670437752078384&theme=dark" width="350" height="500" allowtransparency="true" frameborder="0"></iframe>
